// ainexawebsitenext/ainexawebsitenext/src/utils/Types.ts
export interface ThemeColors {
  BACKGROUND: {
    primary: string;
    secondary: string;
    card: string;
    navbar : string;
    highleter : string;
  };

  TEXT: {
    primary: string;
    secondary: string;
    muted: string;
    inverted: string;
  };

  LINKS: {
    normal: string;
    hover: string;
    visited: string;
  };

  SHADOWS: {
    light: string;
    medium: string;
    heavy: string;
  };

  BORDER: {
    light: string;
    dark: string;
  };

  BUTTON: {
    primary: string;
    primaryText: string;
    secondary: string;
    secondaryText: string;
  };

  STATUS: {
    success: string;
    warning: string;
    error: string;
    info: string;
  };
}